[driver_manager] Add test modules for testing dynamic linking compat driver.
The fake driver entry point has been modified to accept a data array.
The fake driver host will use this to pass the address of the v1 driver
start to the compat driver, which will call the v1 driver start and
return.

This will be used to test the implementation in the next CL.

Bug: 381293474
Change-Id: I6370540118ab8eb7371024c414eb66d73fb9697f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1172099
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>