[fuchsia-component] Avoid opening the instance directory
Rather than open the service instance directory, we can construct an
opener which is capable of opening the service instance member relative
to the service directory. This is more efficient. This change does
require that we wrap the Service object in an Arc as the service proxy
takes a `Box<dyn MemberOpener>` which requires a static lifetime.

This change also changes fidl service proxies to provide a method for
getting the instance name they represent by extending the MemberOpener
trait.

Change-Id: I92d5c5a311b5b5f693cb187a71cd3f8aa134c361
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155213
Reviewed-by: Megan Batty <mgnb@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>
Reviewed-by: Didi She <didis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>