[fidl][rust] Place server and client in dispatcher
Dispatchers need to pass a reference to the server or client to the
handler so it can properly close the connection if an error occurs. This
threads the server or client through so that handlers don't need to
store their own copies at the high level.

Change-Id: Iffe57c7d51b4c27a4f8fc87178ce9f70f405befd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168655
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>