[kernel][arm64] Use hwreg-derived constants in arm64_elX_to_el1()
This change is an incremental step towards migrating the definition of
arm64_elX_to_el1() to libarch for use in physboot, replacing the
hard-coded register #defines with those derived directly from the hwreg
register definitions. There are two functional changes here, both of
which appear to be harmless at this time:
* HCR_EL2 is set before CNTHCTL_EL2, CPTR_EL2 as the layout of the last
  two depend on the first one
* The original routine was setting reserved-as-0 bits in CPTR_EL2
  (OR'ing with 0x33ff instead of 0x2200)

The disassembly before and after was checked locally and these were
indeed the only differences.

Bug: 42164859
Change-Id: If98ad2d07b1129e840001b199c4d870d3a836748
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180439
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>