[kernel][arm64] Reuse sp in arm64_elX_to_el1()
arm64_elX_to_el1() is updated to install the current stack in SP_EL1
before dropping to EL1 so that it may be reused. This should be a no-op
today, but will enable normal use in C++ once this routine is repurposed
for use in physboot.

Bug: 42164859
Change-Id: I262a6091719419d55ecc6005ef8d75912d6fc475
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182356
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>