[bazel] make compdb point to fidling directory
When bazel collects the arguments for c++ files that take a dependency
on a fidl library it uses a path which includes a virtual includes
directory. This directory does not point to a valid location in the
fuchsia tree. For now, we can just point to the GN created fidling
directory so clangd picks it up.

A future change should look into why the virtual includes does not exist
so we can avoid manipulating the arguments more.

Fixes: 382260446
Change-Id: I21d9251e3c5bc05af1935a6f7af5adcb38facbc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182362
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>