[socket-proxy] Wait for response on registry call
After each call to the registry, wait for a response from the server
watcher to ensure that we receive the number of updates we are
expecting.

Move the call that doesn't result in a DNS update to the end so that we
can use `now_or_never`.

Multiply: socket-proxy-integration-test
Fixed: 369741015
Change-Id: I62c87247619d8b27ce892c0ae213c55b3d3563d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126723
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>