[starnix] Don't trust userspace when checking for syscall restarts.
Adds a new field to ThreadState for the current task that we can use to
indicate that a syscall may need to be restarted without having to
scrape the return register for particular errno codes.

There are only three places where we currently mutate the return
register:

1. starnix_core::execution::execute_syscall()
2. starnix_core::signals::signal_handling::prepare_to_restart_syscall()
3. starnix_core::task::do_clone()

(1) is the only place where any of the syscall restart errno codes can
be written. This change updates execute_syscall() to check for
restartable errno codes and stores them in a separate field, ensuring
that all syscalls will still be able to request a restart just by
returning the correct errno.

This new field on ThreadState is then used in
prepare_to_restart_syscall() and must be set to Some(...) in order for
that function to interact at all with user registers, preventing the
kernel from ever attempting to interpret or mutate registers for a
restarting syscall unless it was triggered by our syscall
implementation. Userspace will no longer be able to put the right
values in registers and trap into the kernel with a signal to invoke
the syscall restart logic.

This also prevents the underflow debug assertion we would sometimes see
from signals being delivered at unexpected points in the syscall return
path.

Adds two new tests based on repro snippets in the linked bugs. Confirmed
locally that the crasher fails on x64/arm64/riscv64 without the changes
to Starnix and that the control flow hijack fails on x64 without the
changes to Starnix.

Also includes a minor cleanup to centralize understanding of
syscall-restarting errno codes.

Fixed: 347756382
Fixed: 347063218
Change-Id: Id5489164aadeca82dce5d239550cc3167bf15025
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181025
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>