[kernel][x86] Unify PAGE_SIZE definition
Changes the definition of PAGE_SIZE on x86 to match the other two
architectures. The visible change here is the addition of the `L`
suffix that makes the PAGE_SIZE constant a long instead of int, which
prevents writing code that is typed correctly on one arch and not the
other.

Change-Id: Ica4e3825bd08ee293b75a3da9524d9d3a201696f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177472
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>