[kernel][arm64] Disable EL2 monitor and virtualization tests
Doing so is the plan of record, enabling a more straightforward rollout
for physboot to load the kernel at its ultimate virtual address at EL1.
In particular, this in effect disables our arm64 kernel hypervisor (to
eventually be re-enabled more simply to rely on ARM's EL2 host feature).

This is a simpler reland of fxrev.dev/1176695, which was reverted as a
sequence of changes that seemed to introduce secondary boot failures on
a post-submit builder. More testing was done with this piece to screen
for such failures.

Bug: 42164859
Change-Id: Ib42e423e35793d3305f788c0ba3a4d2e1530cf35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182332
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>