[kernel][phys] Drop to EL1 early in physload
We drop to EL1 - disabling any EL2 monitor - early in phyload, before
setting up the address space. This simplifies the hand-off to EL1, as
then we'll set up the address space in EL1 as opposed to EL2 and don't
have to futz with arm-specific reinstallation. This is prudent also in
that the kernel proper no longer executes in EL2 and we have no need to
do so in physload/physboot.

This is a reland of fxrev.dev/1176696, which was reverted due to
instability around a related change. Nothing has changed with this one.

Bug: 42164859
Change-Id: I036e608a4c8fc6632903df8a78ed778a97847fe1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182457
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>