[kernel][phys][arm64] Hand off address space(s)
As already done with riscv64 and x86, this is a major step toward
physboot properly creating the kernel's address space: it hands off the
address space as it is currently constructed, upper and lower - along
with the kernel's mapping with its final page protections in place -
enabling the kernel to bootstrap from the already-constructed identity
map.

This is a reland of fxrev.dev/1173133, which was reverted due to
instability around secondary boots. It is now believed that this was
due to stores of boot info on the boot CPU not being flushed in time for
loads on secondaries for consultation. This new version addresses that.

Bug: 42164859
Change-Id: Ic59a5f8b57b9a6ef59600bf679d7ac8076d78531
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182458
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>