[bazel] Generate remote_services.bazelrc file.
This is a reland of [1] which was reverted. Compare to patchset 1
on Gerrit to see what was changed.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1177873/7

------ original commit message -------------------------------

Instead of appending its content to bazel-workspace/.bazelrc,
write the content to a separate file, and make the `bazel`
launcher and `bazel_test.py` scripts use it directly.

This also moves the generate_remote_services_bazelrc target
to //build/bazel/bazel_sdk:remote_services_bazelrc, as it is
only used to run the Bazel SDK test suite without generating
anything with the internal Bazel workspace, as it happens
when generating and validating the Code IDK.

Bug: 384956017
Change-Id: If26cd5c9254828d75918a8938b6d5c89ac44d3b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183712
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>