[sestarnix] Fix create_test_executable() to set the context
Fix the create_test_executable() test helper to set the caller-
supplied security context using the "fscreate" mechanism, rather
than by manually setting the xattr, to ensure that the label is set
correctly on the in-memory representation of the node.

Add an explicit test for create_test_executable(), as well as for
the new ScopedFsCreate mechanism, which will be reused in future by
the overlay filesystem labeling hooks.

Bug: 388241218, 364568931
Change-Id: Id305e65de026643df857e364051b4ade11bbc242
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183892
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Przemek Pietrzkiewicz <ppi@google.com>