[fdf][power] Use dynamic offers for power protos
This works around the limitation described in 369189827. We cannot use
dictionaries to provide protocol offers that exist in the static offers
of a cml. This makes the protocols that we want to override into dynamic
offers. This allows us to control whether we offer these from dynamic
offers here, or through a dictionary in the next cl.

Bug: 369189827
Change-Id: I499e73a69cac18de4809d4dd056d3e3965d22ce7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164958
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>