[kernel][vm] Remove extra IPI in ForceFlush on x86
The ForceFlush method in x86 previously sent two IPIs when
invalidating a TLB entry in a restricted page table - the first was
to clear entries in the restricted page table itself, and the second
was to clear entries in the associated unified page table.

This change consolidates the invalidations for both page tables into
1 IPI that causes both PCIDs to be flushed on cores running either
the restricted or unified aspace.

Bug: 319325273
Change-Id: Ifc6eedf90999e070493193e6da833b01c289d972
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177986
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>