[bazel_sdk] created generated_constants file for fuchsia_clang
The generated_constants file needs to be created for the fuchsia_clang
repository. This is done for all code paths except for when we fetch
directly from a cipd tag. This adds the logic to ensure that it is
added.

Change-Id: I01c87b2b1b5a67577feea3ffc8b160775463522d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1184782
Fuchsia-Auto-Submit: Chase Latta <chaselatta@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>