[fxfs] add any range prefix to allocate list immediately
In allocate, if there is a section of the range we need to allocate
before the extent we found, do that right away so all codepaths after
get this case handled for them. Before this, existing overwrite extents
would cause allocate to skip this section of the range.

This also leads to a big simplification to the extent switching range
math that happens later.

The unit tests are expanded to make sure all the allocated ranges
actually end up with overwrite extents covering them, which confirms
there are no other versions of this bug.

Bug: 293943124
Change-Id: I9ad83830024377a47ae8cfdbcbe2cfa6a8124052
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1184618
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>