[eBPF] Type-safe eBPF arguments
Previously it was possible to call `EbpfProgram::run()` with an
arguments of arbitrary types, which may be different from the types
used when the program was verified.  Now it ensures that the passed
argument types match the types that were provided when the program
was verified. Expected argument types are defined in the new
`EbpfProgramContext` trait that replaced `EbpfRunContext`. The
argument types must implement `ProgramArgument` trait which allows
to map Rust types to eBPF types. In most cases types are checked
when a program is linked in `link_program`, except for conformance
tests where the size is not known in compile time. Conformance
tests use `link_program_dynamic` which checks argument type when
`run()` is invoked.

Also `PacketAccessor` trait has been replaced with `Packet` trait
that has to be implemented by the packet type. The interpreter
uses `FromBpfValue` trait to convert `BpfValue` to a packet. This
allows to avoid unsafe type casts outside of the eBPF intepreter.

Bug: 385009220
Change-Id: I745904733954bbd52ab1656306142bdb9e30b844
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177632
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>