[kernel][vm] Introduce FindPageContentLocked
Introduces a FindPageContentLocked method that replaces the primary
usage site of FindInitialPageContentLocked in the LookupCursor. This
new method is more efficient as it allows the first lookup result from
the leaf VMO to be reused instead of being thrown away. The out
parameters are changed to an output struct in preparation for fine
grained locking, which will need additional output information.

The original FindInitialPageContentLocked is re-written to use this new
method, although its functionality is downscoped to just what is needed
by its remaining call sites.

Bug: 338300943
Change-Id: Icaf27bcca044501db1a1922ca9f55c9c48923280
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1157673
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>