[display][coordinator] Clarify test fixture member roles and lifecycles.
This CL renames some data and function members in fixture classes. The
new names are intended to make it easier to understand the returned
values while reading tests.

This CL switches references to/from pointers to match the latest
C++ guidelines. References are guaranteed to be non-null, but must not
be retained. Pointers may be retained, but may be null.

Bug: 388885807
Test: fx test //src/graphics/display
Change-Id: Ibcde59b7ed700dfa1fa2b0bf3bb2cb527c688ab2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185470
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>