[ld][testing] Use explicit ::testing::Sequence in MockLoaderService
The ::testing::InSequence object causes all `EXPECT_CALL(...)`
on *any* mock object to be sequenced with all others.  Instead,
use an explicit ::testing::Sequence object belonging to the
MockLoaderService and pass this in `.InSequence(...)` on every
`EXPECT_CALL(...)` on the mock.  That way, expectations are only
sequenced with respect to other expectations on the same mock.
Since the underlying mock is not exposed by MockLoaderService,
it's easy to ensure this is on all (both) its wrapper methods
that do an actual `EXPECT_CALL(...)`.

Bug: 382340693
Change-Id: I4c9c63c9d380b3448176958118fa9aed24c128db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186315
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>