[power] Add SetBootComplete in the Fake SAG
The Fake SAG used to have logic that prevents setting wrong state. It
is added back and the SetBootComplete API is added in the fake fidl,
so that the fake SAG can pass through the call as well as prevent
wrong state.

Some tests that use the fake SAG are modified.

Bug: b/333699275
Test: CQ
Change-Id: I5822d2d9d669aa44dba87fa032aff816267c55d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162492
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Hai Bi <bihai@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>