[kernel][vm] SetStreamSize unmaps
Calling SetStreamSize unmaps the range from min(new_stream_size,
old_stream_size) to the end of the VMO under the object lock.

This is required to avoid a race when a VMO is mapped with
ZX_VMO_FAULT_BEYOND_STREAM_SIZE and the stream size is reduced. Without
unmapping the pages, it is possible to sneak in a read before the page
is zeroed with ZeroRange.

Run-All-Tests: true
Bug: 379565554
Change-Id: I9804bf0a1a0c434adcbd4ce74f3cd44ed2dd8560
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169435
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Sage Barreda <sagebarreda@google.com>