[audio] ActivityReporter for new usage
Convert protocol ActivityReporter from closed to open, so
as to add a method that clients can use to watch for
activity on AudioRenderUsage2.

In this CL, to minimize churn to clients, the new method is
strict and only render activity is updated. A future CL will
convert the method to flexible and add a companion method
for capture streams.

This is all in service of the new AudioRenderUsage2 enum
value, ACCESSIBILITY.

Bug: b/384972490
Bug: b/384932846
Bug: b/371061877

Change-Id: Icdf8513ab108e309601ce99697a9e01f13548bf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180653
Commit-Queue: Martin Puryear <mpuryear@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>