[audio] UsageReporter for new usage
Convert UsageReporter from closed to open, so as to
add a new watcher method with Usage2 and AudioRenderUsage2.
Also create a new UsageWatcher2 protocol for the method
that notifies clients of usage state changes, since the
caller (AudioCore) must include the new Usage2 as a
parameter.

This is all in service of the new AudioRenderUsage2 enum
value, ACCESSIBILITY.

Bug: b/384973377
Bug: b/371061877

Change-Id: I9703a174b7dc3fc48fed9fbacd76546df09c9e11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180654
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>