[cml] Update duplicate offer message to reflect type of offer
When we have a duplicate offer to "all" with a child component,
the error message should reflect that it's a child component
and not a capability. This previously caused confusion
during a migration where there was both a dictionary
called "diagnostics" as well as a component called "diagnostics".

Fixed: 383624158
Change-Id: I4b3e0768540384a2f7b3c0848cd34e4187c288a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185612
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>