[audio] Update tools for AudioRenderUsage2
Update the audio command-line tools for new render usage
ACCESSIBILITY. This includes audio_listener, wav_player,
signal_generator and audio-capture-timestamp-validator.

Note that all of these are intended to be superceded by
`ffx audio`, but they are updated as part of verifying the
impact of this change on other audio clients.

This is all in service of the new AudioRenderUsage2 enum
value, ACCESSIBILITY.

Bug: b/371061877
Bug: b/384932846

Change-Id: I9f84d28d4bd282690ea99c4419a924b11b175a7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182355
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>