[kernel][vm] Stateless RangeChangeUpdate
Performs a stateless walk of the subtree during a RangeChangeUpdate
instead of building a linked list with state through the children.

In performing a stateless walk some optimization is lost as the range
cannot be perfectly clipped by the children's parent visibility on the
downwalk. In practice this does not matter as the majority of range
change updates are one of:
 * Single page ranges, which can still get perfect clipping due to the
   page list inspection.
 * Large ranges on leaf nodes. These do not have children so there is
   no impact.
 * Large ranges on pager backed root nodes. These nodes do have
   children, but they tend to see the root and not much additional
   invalidation happens as a result of this change.

Aside from the memory savings of removing state from the objects a
stateless traversal is necessary for moving to fine grained locking to
allow multiple concurrent traversals.

Bug: 338300943
Change-Id: I197c320a7d15182fd70bd6cc9fa33f466cfb204e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173412
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>