[assembly][diagnostics] Fix serialization issue with assembly overrides
The serialization format didn't match the original format and was
leading to an error when overrides were used.

The reason this was done is that we can't send objects in structured
config, so the object was being serialized into a string. This assumed
that we never did deserialize(serialize(deserialize(S))), since before
this change serialize(deserialize(S)) != S. However, we do that when
dealing with assembly overrides.

Instead this change guarantees that serialize(deserialize(S)) == S and provides a function to create the string used for the component'
structured config.

The associated bug has the exact repro scenario.

NO_IFTTT=no changes to the the config schema

Fixed: 389747855
Change-Id: I6646766f3283349c86290d7d30d59d94012f1347
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187292
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>