[audio] Rebaseline before FIDL flexibility changes
A collection of small unrelated changes, to "clear the
water" so that subsequent layers of changes are more clear.

This includes simplifying the 'ffx audio' protocol for play
as well as various small comment fixes.

This is all in service of the new AudioRenderUsage2 enum
value, ACCESSIBILITY.

Bug: b/371061877
Bug: b/384932846
Change-Id: I2eda54f01c4f604c5fbe166622d96a05b4b3d261
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186519
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>