[starnix][cgroup] add_process takes TempRef<ThreadGroup> instead of WeakRef
We need TempRef to freeze the tasks of a newly added process if the
cgroup is frozen. We already upgrade the ThreadGroup to check that it is
alive, but previously we downgrade then upgrade again if we need to
freeze it. Pass in a TempRef directly to avoid the unnecessary
downgrade/upgrade.

Also remove freeze_process as it is no longer needed.

Bug: 388081859
Change-Id: I99279199a71939d29e7b2531d17ca094b92f2171
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187812
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Song <wintermelons@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>