[tee][runtime] Hang storage objects off of context
This adds a 'storage' entry to the api_impl::Context struct and hangs
the persistent object tracking logic off of the context. This
consolidates static state in the runtime and prepares for implementing
transient objects and operations that have to interact with persistent
and/or transient objects.

We may also want to replace some of the runtime mutability with static
mutability checking inside the persistent object management. This defers
on any changes in that direction to keep the changes simpler.

Bug: 360942417

Change-Id: I6aa3fcd8b52da3132b13bf750c81298c8fbe6bfb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185927
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>