[utest] Use relative offset for mapping
Restricted mode testing for aarch32 attempts to keep all memory shared
with the restricted environment under the addressable limit for that
code.  That limit is expressed as an absolute, but zx_vmar_map() expects
the offset to be relative to the base of the VMAR.

This error manifested as test flakiness because the memory was usually
mapped within the absolute range.  This change fixes the flakiness and
ensures any shared memory is mapped under the provided limit.

Bug: 389697363
Change-Id: Ia3b3cb19c00cbd29e1864535bf8cd5441fa44be7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187672
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Will Drewry <drewry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>