[eBPF] Read packet length from the context struct
Previously cBPF programs were receiving packet length as a second
argument and that value was used by the BPF_LEN implementation. That
argument is redundant: in socket filters the value is stored in the
`sk_buff` struct passed as the first argument, while it's constant in
seccomp programs.

This change adds `CbpfConfig`, which allows to customize cBPF
converter behavior. This allows to use different behavior for
socket filters and seecomp programs. The same config also allows to
disable the `BPF_MSH` instruction in seccomp as it doesn't make
sense there.

Change-Id: I1e6e10b285eba62080759823cccfed7191970f53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185919
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>