[power] Prepare the move of shutdown_watcher
This file is a copy from the src/power/power-manager/src. It is not
being used yet after this lands, until the next CL. In the next CL,
this copy of shutdown_watcher will be modified and the original will
be removed. In this way, gerrit will regard the file as a change
next time instead of as a new file.

Bug: b/366035937
Test: None
Change-Id: I02f1ec5744960f71c2b2aae60bae2e6f71f8eab2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188193
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Hai Bi <bihai@google.com>