[zircon][vm] Insert new VmCowPages nodes as hidden
When creating the first bidirectional clone of a node in the VmCowPages
tree, we need to split the current node into a hidden parent, which
holds all the pages, and a child that will hold new, dirty pages.
Previously, we created an empty node for the child and mutated the
current node to make it hidden. After this CL, we create the hidden
parent as a fresh VmCowPages object and move all the pages from the
existing node to this parent.

This CL is expected to regress a number of VMO clone microbenchmarks.
These regressions are necessary in order to restructure the VmCowPages
data structure to unlock larger performance wins in future CLs.

Bug: b/338300943
Change-Id: Icdf20f19f8651fa6335f274c25924f9edaca04e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186497
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>