[self_profile_report] Summarize non-leaves in custom rollups
Change the behavior of custom rollups to cover non-leaf nodes because
limiting custom rollups to leaf nodes seems unnecessary and misleading,
for example when the same function is called both as a leaf node and
non-leaf node.

With this change, the customized rollups in the self-profile reports
only depend on name-prefixes specified by customers, which seems to
be flexible and straightforward enough.

Bug: b/389923307

Change-Id: Iff666404e479d4b20a3eb1e85c9b22c033aaa812
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188614
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Ali Zhang <alizhang@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Ali Zhang <alizhang@google.com>