[kernel][vm] Limit request dumping
Limits the amount of information emitted to the debuglog during a page
request overtime event by capping the number of items in lists that are
printed, and not redundantly dumping the source/provider state for
every waiting request.

Without this change, if there are multiple requests on a pager, then
each request prints out the information about every other request. This
causes the amount of information to be quadratic in the number of
requests, with most of it being duplicated and much of the remainder
being unhelpful.

Bug: 381312579
Change-Id: I00b9f5f4192dd73701d9f74f8467c761f610fe45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186933
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>