[acpi] Don't prematurely encode the bus metadata
Make the acpi driver not encode the bus metadata before passing it to
the device that will serve the bus metadata. Instead, the device will
perform the encoding.

In a future CL, the device will serve the bus metadata with
`fdf_metadata::MetadataServer` which will require non-encoded metadata
in order to initialize.

Change-Id: I012eb13692b893b2b6d248e58558c371af731e22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181692
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>