[ffx]: add structure to top level ffx errors
Introduces new type SerializableError and implements serde::Serialize for it.

Then the ffx_command::exit function now inspects the command line to see
if the machine writer was requested. If so it serializes the Error
variant to JSON and writes that to stderr. Otherwise, it just prints the
error string.

We need this additional type as we cannot directly serialize the
FfxError or ffx_command::Error since they have an `anyhow::Error`
embedded in them (which cannot be serialized).

Change-Id: Ia3007d28c4251a28d8cacb1f6772e9106d3117c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1161889
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>