[bazel_sdk][tests] Fix bazel_test.py for the in-tree @fuchsia_sdk
This CL changes the implementation and interface of the
bazel_test.py script to ensure that it correctly runs against
the in-tree @fuchsia_sdk and @fuchsia_in_tree_idk repositories,
when invoked with the new `--fuchsia-in-tree-sdk` flag.

The `--fuchsia_sdk_dir` option is renamed to `--fuchsia_sdk_directory`
which is consistent with `--fuchsia_idk_directory`. For now it still
expects a self-contained SDK directory (i.e. one that does not load
rules from @rules_fuchsia).

+ Update README.md accordingly.

Fixed: 383536158
Change-Id: Ib8e35b815102d9aea99cb6a12fcde2df8afb3cf4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171837
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>