[netpol] Add CheckPresence to SocketProxy protocol
SocketProxy will be an optional component determined on a per-product
basis. To determine whether SocketProxy is available in the product, a
FIDL call must be made against the protocol to determine whether the
connection is valid.

The future client of FuchsiaNetworks, netcfg, uses this approach to
determine which filtering and DHCP protocol to use.

The existing endpoints offered by the API have side-effects (such as
writing to Inspect) so a new endpoint should exist for this behavior.

Bug: 385368895
Test: Verified manually on https://fxrev.dev/1165212 with use of
CheckPresence

Change-Id: I4ac78e3176f2f085d40f9bec7cc76f05a8b01f39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186496
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>