[component] Simplify log messages
Too many monikers being logged! Not every level of logging needs to
log the moniker of the component with an error. It's sufficient to log
the moniker only when there are multiple paths into the error (and
when that error might not be wrapped by another error that also logs
the moniker).

Bug: b/384927637
Change-Id: I5d84521e4b77c7af3c6d37b6acfda61a74814373
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185554
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>