[starnix] Refactor container namespace proxying
The ultimate goal of this change is to allow remotefs to take in
arbitrary paths, such as /userdata/data. If the root directory is a
valid namespace, then we'll create necessary subdirs (e.g. data) and
return them as the handle, per the existing contract.

Since we need 1..N access to arbitrary entries in the namespace, this
change refactors the whole of how the container passes along namespace
entries. Rather than transforming and relinquishing namespace handles
directly, it supplies the kernel with cloneable proxy references which
it can transform and dish out as needed.

In order to soft transition the existing remotefs creation (which only
mounts the entire source as `.`), I've placed logic within the remotefs
creation to assume the data directory. In a subsequent change, I'll be
migrating the fstab to specify the correct namespace and subdirs
explicitly, and then remove that logic.

Bug: 379929394
Test: syscall mount_test, starnix-tests, manually verified boot process.
Change-Id: I7fbc766938602ef5fcec63493ebd4988f9117efc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180339
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>